+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/components/ActivePortalAppHost.tsx
@@ -1,18 +1,34 @@
-import React, { useRef, useCallback, useEffect } from 'react';
+import React, {useRef, useCallback, useEffect, useMemo} from 'react';
+import {nanoid} from 'nanoid';
 import { CircularProgress } from '@mashroom/mashroom-portal-ui-commons';
-import {setHostWidth as setHostWidthAction} from '../store/actions';
+import {addMessagePublishedByApp, setHostWidth as setHostWidthAction, setTopicsSubscribedByApp} from '../store/actions';
 import useStore from '../store/useStore';
+import loadPortalApp from '../load-portal-app';
+import unloadPortalApp from '../unload-portal-app';
+import getMessageBusPortalAppUnderTest from '../message-bus-portal-app-under-test';
+import type {LoadedPortalApp} from '../types';
 
-type Props = {
-    hostElementId: string;
-}
-
-export default ({hostElementId}: Props) => {
-    const wrapperElemRef = useRef<HTMLDivElement | null>(null);
+export default () => {
     const activePortalApp = useStore((state) => state.activePortalApp);
     const {width: hostWidth} = useStore((state) => state.host);
     const dispatch = useStore((state) => state.dispatch);
     const setHostWidth = (hostWidth: string) => dispatch(setHostWidthAction(hostWidth));
+    const hostElementId = useMemo(() => `mashroom-sandbox-app-host-elem_${nanoid(8)}`, []);
+    const hiddenHostElementId = useMemo(() => `mashroom-sandbox-app-host-elem_${nanoid(8)}`, []);
+    const messageBusPortalAppUnderTest = useMemo(() => {
+        const messageBus = getMessageBusPortalAppUnderTest();
+        messageBus.onMessageSent((topic, data) => {
+            dispatch(addMessagePublishedByApp({
+                topic,
+                data
+            }));
+        });
+        messageBus.onTopicsChanged((topics) => {
+            dispatch(setTopicsSubscribedByApp(topics));
+        });
+        return messageBus;
+    }, [activePortalApp]);
+    const wrapperElemRef = useRef<HTMLDivElement | null>(null);
 
     const resizerMouseMove = useCallback((event: MouseEvent) => {
         if (!wrapperElemRef.current) {
@@ -42,12 +58,31 @@ export default ({hostElementId}: Props) => {
         };
     }, [resizerMouseMove, resizerMouseUp]);
 
+    useEffect(() => {
+        if (activePortalApp) {
+            const { appName, setup } = activePortalApp!;
+            let loadedApp: LoadedPortalApp | undefined;
+            (async () => {
+                try {
+                    loadedApp = await loadPortalApp(appName, hostElementId, hiddenHostElementId, setup, messageBusPortalAppUnderTest);
+                } catch (error) {
+                    console.error('Loading app failed', error);
+                }
+            })();
+            return () => {
+                if (loadedApp) {
+                    unloadPortalApp(loadedApp);
+                }
+            };
+        }
+    }, [activePortalApp]);
+
     if (!activePortalApp) {
         return null;
     }
 
     let currentWidth = hostWidth;
-    // Ensure width is in 'px' if it's just a number string
+    // Ensure the width is in 'px' if it's just a number string
     if (currentWidth.match(/^\d+$/)) {
         currentWidth = `${currentWidth}px`;
     }
@@ -76,6 +111,7 @@ export default ({hostElementId}: Props) => {
             >
                 <div className='grip' />
             </div>
+            <div id={hiddenHostElementId} style={{ display: 'none' }} />
         </div>
     );
 };