+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/components/PortalApp.tsx
@@ -1,14 +1,5 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import {useTranslation} from 'react-i18next';
-import loadPortalApp from '../load-portal-app';
-import { mergeAppConfig } from '../utils';
-import {
-    QUERY_PARAM_APP_NAME,
-    QUERY_PARAM_APP_CONFIG,
-    QUERY_PARAM_PERMISSIONS,
-    QUERY_PARAM_LANG,
-    QUERY_PARAM_WIDTH
-} from '../constants';
+import React, { useEffect, useCallback } from 'react';
+import {mergeAppConfig} from '../utils';
 import {
     setActivePortalApp as setActivePortalAppAction,
     setAppLoadingError as setAppLoadingErrorAction,
@@ -18,9 +9,10 @@ import {
 } from '../store/actions';
 import useStore from '../store/useStore';
 import PortalAppSelection from './PortalAppSelection';
-import PortalAppConfig from './PortalAppConfig';
+import SelectedPortalAppConfig from './SelectedPortalAppConfig';
+import ActivePortalAppDescription from './ActivePortalAppDescription';
 
-import type {PortalAppQueryParams, ActivePortalApp, SelectedPortalApp, MessageBusPortalAppUnderTest} from '../types';
+import type {PortalAppQueryParams, ActivePortalApp, SelectedPortalApp} from '../types';
 import type {
     MashroomKnownPortalApp,
     MashroomPortalAppService,
@@ -28,18 +20,13 @@ import type {
 } from '@mashroom/mashroom-portal/type-definitions';
 
 type Props = {
-    hostElementId: string;
     queryParams: PortalAppQueryParams;
     portalAppService: MashroomPortalAppService;
     messageBus: MashroomPortalMessageBus;
-    messageBusPortalAppUnderTest: MessageBusPortalAppUnderTest;
 }
 
-export default ({hostElementId, queryParams, portalAppService, messageBus, messageBusPortalAppUnderTest}: Props) => {
-    const {t} = useTranslation();
-    const [permalinkCopiedToClipboard, setPermalinkCopiedToClipboard] = useState<boolean>(false);
+export default ({queryParams, portalAppService, messageBus}: Props) => {
     const activePortalApp = useStore((state) => state.activePortalApp);
-    const {width: currentHostWidth} = useStore((state) => state.host);
     const dispatch = useStore((state) => state.dispatch);
     const setKnownApps = (availableApps: Array<MashroomKnownPortalApp>) => dispatch(setKnownAppsAction(availableApps));
     const setSelectedPortalApp = (app: SelectedPortalApp | undefined | null) => dispatch(setSelectedPortalAppAction(app));
@@ -49,21 +36,11 @@ export default ({hostElementId, queryParams, portalAppService, messageBus, messa
 
     const loadPortalAppInternal = useCallback(async (selectedPortalApp: SelectedPortalApp, hostWidth: string | undefined | null): Promise<void> => {
         console.info(`Loading app '${selectedPortalApp.appName}' with setup: `, selectedPortalApp.setup);
-
-        const {appName, setup} = selectedPortalApp;
         setActivePortalApp(selectedPortalApp);
-
         if (hostWidth) {
             setHostWidth(hostWidth);
         }
-
-        try {
-            await loadPortalApp(appName, hostElementId, setup, messageBusPortalAppUnderTest);
-        } catch (error) {
-            console.error('Loading app failed', error);
-        }
-    }, [hostElementId, messageBusPortalAppUnderTest]);
-
+    }, []);
 
     const selectionChanged = useCallback(async (appName: string | undefined | null): Promise<void> => {
         setTimeout(() => {
@@ -98,7 +75,7 @@ export default ({hostElementId, queryParams, portalAppService, messageBus, messa
     }, [queryParams, messageBus]);
 
     useEffect(() => {
-        const loadApps = async () => {
+        (async () => {
             try {
                 const apps = await portalAppService.searchApps({includeNotPermitted: true});
                 setKnownApps((apps || []).sort((app1, app2) => app1.name.localeCompare(app2.name)));
@@ -109,70 +86,20 @@ export default ({hostElementId, queryParams, portalAppService, messageBus, messa
             } catch (error) {
                 console.error('Loading available apps failed', error);
             }
-        };
-
-        loadApps();
+        })();
     }, [queryParams.appName, queryParams.preselectAppName, selectionChanged]);
 
-    const copyToClipboard = (text: string) => {
-        navigator.clipboard.writeText(text);
-        setPermalinkCopiedToClipboard(true);
-        // Optionally, reset the message after a few seconds
-        setTimeout(() => setPermalinkCopiedToClipboard(false), 3000);
-    };
 
     let content;
-
     if (activePortalApp) {
-        const queryParamsArray: Array<string> = [];
-        queryParamsArray.push(`${QUERY_PARAM_APP_NAME}=${encodeURIComponent(activePortalApp.appName)}`);
-        queryParamsArray.push(`${QUERY_PARAM_APP_CONFIG}=${btoa(JSON.stringify(activePortalApp.setup.appConfig))}`);
-        queryParamsArray.push(`${QUERY_PARAM_PERMISSIONS}=${btoa(JSON.stringify(activePortalApp.setup.user.permissions))}`);
-        queryParamsArray.push(`${QUERY_PARAM_LANG}=${activePortalApp.setup.lang}`);
-        queryParamsArray.push(`${QUERY_PARAM_WIDTH}=${encodeURIComponent(currentHostWidth)}`);
-        const permalink = `${document.location.origin + document.location.pathname}?${queryParamsArray.join('&')}`;
-
         content = (
-            <>
-                <div className='mashroom-sandbox-app-output-row'>
-                    <div>
-                        {t('sandboxPermalink')}
-                    </div>
-                    <div>
-                        <a href={permalink} target='_blank' rel="noreferrer">
-                            {t('link')}
-                        </a>
-                        <div className='mashroom-sandbox-copy-permalink' onClick={() => copyToClipboard(permalink)}></div>
-                        {permalinkCopiedToClipboard && (
-                            <div className='mashroom-sandbox-permalink-copied'>
-                                {t('copiedToClipboard')}
-                            </div>
-                        )}
-                    </div>
-                </div>
-                <div className='mashroom-sandbox-app-output-row'>
-                    <div>
-                        {t('appName')}
-                    </div>
-                    <div>
-                        <strong id='mashroom-sandbox-app-name'>{activePortalApp.appName}</strong>
-                    </div>
-                </div>
-                <div className='mashroom-sandbox-app-output-row'>
-                    <div>
-                        {t('appSetup')}
-                    </div>
-                    <div>
-                        <pre id='mashroom-sandbox-app-setup'>{JSON.stringify(activePortalApp.setup, null, 2)}</pre>
-                    </div>
-                </div>
-            </>
+            <ActivePortalAppDescription />
         );
     } else {
         content = (
             <>
                 <PortalAppSelection preselectAppName={queryParams.preselectAppName} onSelectionChanged={selectionChanged} />
-                <PortalAppConfig sbAutoTest={queryParams.autoTest} onConfigSubmit={loadPortalAppInternal} />
+                <SelectedPortalAppConfig sbAutoTest={queryParams.autoTest} onConfigSubmit={loadPortalAppInternal} />
             </>
         );
     }