+++ packages/plugin-packages/mashroom-portal/type-definitions/api.ts
@@ -575,7 +575,7 @@ export type MashroomPortalPageEnhancement = {
      */
     readonly lastReloadTs: number;
     /**
-     * Resources root URI (can be file, http or https)
+     * Resources root URI (can be a file, http or https)
      */
     readonly resourcesRootUri: string;
     /**
@@ -683,7 +683,7 @@ export interface MashroomPortalService {
     getSite(siteId: string): Promise<MashroomPortalSite | null | undefined>;
 
     /**
-     * Find the site with given path
+     * Find the site with the given path
      */
     findSiteByPath(path: string): Promise<MashroomPortalSite | null | undefined>;
 
@@ -708,7 +708,7 @@ export interface MashroomPortalService {
     getPage(pageId: string): Promise<MashroomPortalPage | null | undefined>;
 
     /**
-     * Find the page ref within a site with given friendly URL
+     * Find the page ref within a site with a given friendly URL
      */
     findPageRefByFriendlyUrl(site: MashroomPortalSite, friendlyUrl: string): Promise<MashroomPortalPageRef | null | undefined>;
 
@@ -718,7 +718,7 @@ export interface MashroomPortalService {
     findPageRefByPageId(site: MashroomPortalSite, pageId: string): Promise<MashroomPortalPageRef | null | undefined>;
 
     /**
-     * Insert new page
+     * Insert a new page
      */
     insertPage(page: MashroomPortalPage): Promise<void>;
 
@@ -728,7 +728,7 @@ export interface MashroomPortalService {
     updatePage(page: MashroomPortalPage): Promise<void>;
 
     /**
-     * Insert new page
+     * Insert a new page
      */
     deletePage(req: Request, pageId: string): Promise<void>;
 
@@ -763,6 +763,12 @@ export type AppSearchFilter = {
 
 export type ModalAppCloseCallback = () => void;
 
+export type ClientBootstrapReference = {
+    readonly pluginName: string;
+    readonly instanceId: string;
+    clientBootstrap: MashroomPortalAppPluginBootstrapFunction;
+}
+
 export interface MashroomPortalAppService {
     /**
      * Get all Portal Apps available to the user
@@ -775,24 +781,24 @@ export interface MashroomPortalAppService {
     searchApps(filter?: AppSearchFilter): Promise<Array<MashroomKnownPortalApp>>;
 
     /**
-     * Load a Portal App into given host element at given position (or at the end if position is not set)
+     * Load a Portal App into a given host element at a given position (or at the end if the position is not set)
      *
-     * The returned promise will always resolve! If there was a loading error the MashroomPortalLoadedPortalApp.error property will be true.
+     * The returned promise will always be resolved! If there was a loading error, the MashroomPortalLoadedPortalApp.error property will be true.
      */
     loadApp(appAreaId: string, pluginName: string, instanceId: string | null | undefined, position?: number | null | undefined, overrideAppConfig?: any | null | undefined): Promise<MashroomPortalLoadedPortalApp>;
 
     /**
      * Load a Portal App into a modal overlay.
      *
-     * The returned promise will always resolve! If there was a loading error the MashroomPortalLoadedPortalApp.error property will be true.
+     * The returned promise will always be resolved! If there was a loading error, the MashroomPortalLoadedPortalApp.error property will be true.
      */
     loadAppModal(pluginName: string, title?: string | null | undefined, overrideAppConfig?: any | null | undefined, onClose?: ModalAppCloseCallback | null | undefined): Promise<MashroomPortalLoadedPortalApp>;
 
     /**
-     * Reload given Portal App.
+     * Reload the given Portal App.
      *
-     * The returned promise will always resolve!
-     * If there was a loading error the MashroomPortalLoadedPortalApp.error property will be true.
+     * The returned promise will always be resolved!
+     * If there was a loading error, the MashroomPortalLoadedPortalApp.error property will be true.
      */
     reloadApp(id: string, overrideAppConfig?: any | null | undefined): Promise<MashroomPortalLoadedPortalApp>;
 
@@ -805,6 +811,14 @@ export interface MashroomPortalAppService {
      * Move a loaded App to another area (to another host element within the DOM)
      */
     moveApp(id: string, newAppAreaId: string, newPosition?: number): void;
+
+    /**
+     * Loads the Portal App without starting it and returns a reference to the client bootstrap.
+     *
+     * ONLY use this if you exactly know what you are doing!
+     * If you start the Portal App, you have to take care of calling the lifecycle methods yourself.
+     */
+    loadAppClientBootstrap(hostElementId: string, pluginName: string): Promise<ClientBootstrapReference>;
 
     /**
      * Show the name and version for all currently loaded App in an overlay (for debug purposes)
@@ -827,7 +841,7 @@ export interface MashroomPortalAppService {
     unregisterAppLoadedListener(listener: MashroomPortalAppLoadListener): void;
 
     /**
-     * Add a listener for unload events (fired before an App will be detached from the page)
+     * Add a listener for unloaded events (fired before an App will be detached from the page)
      */
     registerAppAboutToUnloadListener(listener: MashroomPortalAppLoadListener): void;
 
@@ -837,7 +851,7 @@ export interface MashroomPortalAppService {
     unregisterAppAboutToUnloadListener(listener: MashroomPortalAppLoadListener): void;
 
     /**
-     * Load the setup for given App/plugin name on the current page
+     * Load the setup for a given App / plugin name on the current page
      */
     loadAppSetup(pluginName: string, instanceId: string | null | undefined): Promise<MashroomPortalAppSetup>;
 
@@ -855,7 +869,7 @@ export interface MashroomPortalAppService {
     checkLoadedPortalAppsUpdated(): Promise<Array<string>>;
 
     /**
-     * Prefetch resources of given App/plugin. This is useful if you know which apps you will have to load
+     * Prefetch resources of a given App / plugin. This is useful if you know which apps you will have to load
      * in the future and want to minimize the loading time.
      */
     prefetchResources(pluginName: string): Promise<void>;
@@ -875,7 +889,7 @@ export interface MashroomPortalSiteService {
     getSites(): Promise<Array<MashroomPortalSiteLinkLocalized>>;
 
     /**
-     * Get the page tree for given site
+     * Get the page tree for a given site
      */
     getPageTree(siteId: string): Promise<Array<MashroomPortalPageRefLocalized>>;
 }
@@ -886,7 +900,7 @@ export interface MashroomPortalPageService {
      */
     getCurrentPageId(): string;
     /**
-     * Get the page friendlyUrl from given URL (e.g. /portal/web/test?x=1 -> /test)
+     * Get the page friendlyUrl from the given URL (e.g., /portal/web/test?x=1 -> /test)
      */
     getPageFriendlyUrl(pageUrl: string): string;
     /**
@@ -894,7 +908,7 @@ export interface MashroomPortalPageService {
      */
     getPageId(pageUrl: string): Promise<string | undefined>;
     /**
-     * Get the content for given pageId.
+     * Get the content for the given pageId.
      * It also calculates if the correct theme and all necessary page enhancements for the requested page
      * are already loaded. Otherwise, fullPageLoadRequired is going to be true and no content returned.
      */
@@ -963,7 +977,7 @@ export interface MashroomPortalAdminService {
     getExistingRoles(): Promise<Array<RoleDefinition>>;
 
     /**
-     * Get all app instances on current page
+     * Get all app instances on the current page
      */
     getAppInstances(): Promise<Array<MashroomPagePortalAppInstance>>;
 
@@ -1003,7 +1017,7 @@ export interface MashroomPortalAdminService {
     getPage(pageId: string): Promise<MashroomPortalPage>;
 
     /**
-     * Add new page
+     * Add a new page
      */
     addPage(page: MashroomPortalPage): Promise<MashroomPortalPage>;
 
@@ -1038,7 +1052,7 @@ export interface MashroomPortalAdminService {
     getSite(siteId: string): Promise<MashroomPortalSite>;
 
     /**
-     * Add new site
+     * Add a new site
      */
     addSite(site: MashroomPortalSite): Promise<MashroomPortalSite>;
 
@@ -1069,7 +1083,7 @@ export type MashroomPortalMessageBusInterceptor = (data: any, topic: string, sen
 
 export interface MashroomPortalMessageBus {
     /**
-     * Subscribe to given topic.
+     * Subscribe to a given topic.
      * Topics starting with getRemotePrefix() will be subscribed server side via WebSocket (if available).
      * Remote topics can also contain wildcards: # for multiple levels and + or * for a single level
      * (e.g. remote:/foo/+/bar)
@@ -1077,19 +1091,19 @@ export interface MashroomPortalMessageBus {
     subscribe(topic: string, callback: MashroomPortalMessageBusSubscriberCallback): Promise<void>;
 
     /**
-     * Subscribe once to given topic. The handler will be removed after the first message has been received.
+     * Subscribe once to the given topic. The handler will be removed after the first message has been received.
      * Remote topics are accepted.
      */
     subscribeOnce(topic: string, callback: MashroomPortalMessageBusSubscriberCallback): Promise<void>;
 
     /**
-     * Unsubscribe from given topic.
+     * Unsubscribe from the given topic.
      * Remote topics are accepted.
      */
     unsubscribe(topic: string, callback: MashroomPortalMessageBusSubscriberCallback): Promise<void>;
 
     /**
-     * Publish to given topic.
+     * Publish to the given topic.
      * Remote topics are accepted.
      */
     publish(topic: string, data: any): Promise<void>;
@@ -1108,7 +1122,7 @@ export interface MashroomPortalMessageBus {
     /**
      * Register a message interceptor.
      * An interceptor can be useful for debugging or to manipulate the messages.
-     * It can change the data of an event by return a different value or block messages
+     * It can change the data of an event by returning a different value or block messages
      * by calling cancelMessage() from the interceptor arguments.
      */
     registerMessageInterceptor(interceptor: MashroomPortalMessageBusInterceptor): void;
@@ -1122,36 +1136,36 @@ export interface MashroomPortalMessageBus {
 export interface MashroomPortalMasterMessageBus
     extends MashroomPortalMessageBus {
     /**
-     * Get an App specific instance.
+     * Get an App-specific instance.
      * The returned instance will set the senderId on the MashroomPortalMessageBusSubscriberCallback to the given id.
      */
     getAppInstance(appId: string): MashroomPortalMessageBus;
 
     /**
-     * Unsubscribe/Unregister everything from a given App (for cleanup after unload)
+     * Unsubscribe/Unregister everything from a given App (for cleanup after unloading)
      */
     unsubscribeEverythingFromApp(appId: string): Promise<void>;
 }
 
 export interface MashroomPortalStateService {
     /**
-     * Get a property from state.
+     * Get a property from the state.
      * It will be looked up in the URL (query param or encoded) and in the local and session storage
      */
     getStateProperty(key: string): any | null | undefined;
 
     /**
-     * Add given key value pair into the URL (encoded)
+     * Add the given key value pair into the URL (encoded)
      */
     setUrlStateProperty(key: string, value: any | null | undefined): void;
 
     /**
-     * Add given key value pair to the session storage
+     * Add the given key value pair to the session storage
      */
     setSessionStateProperty(key: string, value: any): void;
 
     /**
-     * Add given key value pair to the local storage
+     * Add the given key value pair to the local storage
      */
     setLocalStoreStateProperty(key: string, value: any): void;
 }
@@ -1197,7 +1211,7 @@ export interface MasterMashroomPortalRemoteLogger extends MashroomPortalRemoteLo
     info(msg: string, portalAppName?: string | null | undefined): void;
 
     /**
-     * Get an app specific instance.
+     * Get an App-specific instance.
      * The returned instance will set the portalAppName automatically.
      */
     getAppInstance(portalAppName: string): MashroomPortalRemoteLogger;
@@ -1221,13 +1235,13 @@ export type MashroomPortalClientServices = {
 
 export type MashroomPortalAppLifecycleHooks = {
     /**
-     * Will be called before the host element will be removed from the DOM.
-     * Can be used to clean up (e.g. to unmount a React App).
+     * Will be called before the host element is removed from the DOM.
+     * Can be used to clean up (e.g., to unmount a React App).
      */
     readonly willBeRemoved?: () => void | Promise<void>;
     /**
      * Dynamically update the appConfig.
-     * If present this will be used to update the appConfig instead of restarting the whole App.
+     * If present, this will be used to update the appConfig instead of restarting the whole App.
      */
     readonly updateAppConfig?: (appConfig: MashroomPluginConfig) => void;
 };
@@ -1242,7 +1256,7 @@ export type MashroomPortalAppSSRRemoteRequest = {
 
 export type MashroomPortalAppSSRResultEmbeddedApp = {
     /**
-     * The area Id (host ID) the embedded Portal App should be integrated into
+     * The area id (host ID) the embedded Portal App should be integrated into
      */
     readonly appAreaId: string;
     /**